backend/requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
pip==21.2.3
setuptools==57.4.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3